Snapshot: don't fail on unparseable symbols. (#118)
* Snapshot: don't fail on unparseable symbols.

Previously, the `scip snapshot` command failed if there was a symbol
that wasn't parseable. This was unhelpful when trying to debug why
navigation was not working as expected (which is the main situation
where I use `scip snapshot`). Unparseable symbols are not invalid,
they just don't work with cross-repo navigation. To fail fast
like before, there's a new `scip snapshot --strict` flag that
fails if there's a parser error.

* Enable strict by default and explain how to disable strict