Remove debugging output from symbol.rs that cannot otherwise be disabled. (#130)
* Remove use of dbg! from symbol.rs.

As documented at https://doc.rust-lang.org/std/macro.dbg.html the dbg!
macro unconditionally prints its argument to stderr, even in release
builds.  (It is not possible to quiet it through use of RUST_LOG.)

Presumably this was accidentally left in.

* Remove unconditional use of println in symbol.rs.

This unconditional debugging information probably also was not
intended.