Add `SymbolInformation.Kind`
This is the first step towards closing https://github.com/sourcegraph/scip/issues/154.

Previously, the only way to determine the type of a symbol was through `SymbolDescriptor.Suffix`. This solution was not ideal since `Suffix` had very coarse grained types, and it was intentionally named "suffix" to indicate it was primarily used to determine the syntax of the symbol (not semantics).

This PR adds a new enum `SymbolInformation.Kind` to add first-class support to determine the type of a symbol.