Add `SymbolInformation.display_name`
Towards https://github.com/sourcegraph/scip/issues/154

Previously, there was no reliable way to get the name of a symbol. You
could parse it from the symbol descriptor but that only worked for
global symbols (not locals) and was error-prone when the indexer chooses
to a special encoding in the symbol for some reason (for example,
all-lowercase because symbols are case-insensitive).

Now, clients can use `SymbolInformation.display_name` to reliably get
the name of a symbol to render in API docs or some other application.