Add `SymbolInformation.enclosing_symbol`
Towards https://github.com/sourcegraph/scip/issues/154

Previously, it was not possible to determine the enclosing symbol of a
local because the syntax for local symbols has no hierarchy (unlike
global symbols). This PR closes the gap by adding a field
`SymbolInformation.enclosing_symbol` to allow documenting the "parent"
or "owner" of a local symbol.