feat: Speed up symbol parsing by minimizing allocations (#258)
Changes the symbol parsing logic to minimize allocations. In particular,
when we only care about validating symbols (e.g. during document
canonicalization when ingesting uploads), there is really ~no need to
allocate any strings at all. Validation and parsing share most of the
underlying code -- the only change is we create "writer" types which
will discard writes (and hence any internal buffer growth) when we're
only in validation mode.

For parsing mode, it is now possible to pass in pre-allocated values
which will be overwritten as possible (unless the symbol is very large,
in which case it still needs to allocate new Descriptors).

See PR for benchmarks.